Refactor logger imports for consistency and simplicity.
Replaced relative imports of the logger module to ensure consistent usage. This change simplifies the import structure and avoids potential pitfalls with deeply nested relative paths. It improves maintainability across all affected files.
