+++ services/telegram_service/app/flow_integration.py
@@ -12,7 +12,7 @@ from common.messaging.unified_flow import (
 )
 
 # Import service logger and logging utilities
-from .. import logger
+from . import logger
 from common.utils.logging_config import log_operation, log_context
