+++ services/telegram_service/app/handlers/menu_handlers.py
@@ -26,7 +26,7 @@ from ..keyboards import (
 )
 
 # Import service logger and logging utilities
-from ... import logger
+from .. import logger
 from common.utils.logging_config import log_operation, log_context
