+++ services/telegram_service/app/handlers/phone_verification.py
@@ -21,7 +21,7 @@ from ..keyboards import (
 )
 
 # Import service logger and logging utilities
-from ... import logger
+from .. import logger
 from common.utils.logging_config import log_operation, log_context
