+++ services/telegram_service/app/utils/message_utils.py
@@ -12,7 +12,7 @@ from common.messaging.unified_platform_utils import (
 )
 
 # Import service logger and logging utilities
-from ... import logger
+from .. import logger
 from common.utils.logging_config import log_operation, log_context
